removed paths from tsconfig
